refactor: api 분리 및 모듈화
